Addressed comments
Moved the data into a data folder in the src folder. changed the casing of the variables to be consistent with first letter lower case camel casing. instead of moving to constructor made the main method into an initiate method that can be called from the main class file once we make it. Since the initiator makes an instance of gui I couldn't put it into the constructor.