Update GUI.java
Also added the ability in the UI for a group size to be specified between the sizes of 3-6. Any less than 3 and married couples couldn't be in a small group, and any more and the smallest people data would only turn up 3-2 groups per itteration, which seemed kinda low for this kind of system.