Updated GUI, Main, and pom so the program is now runnable
GUI is updated to include the option for a custom file, and now buttons perform most of the function when executed, once the algorithm is finished in the group file it can be completed.

Main was updated to run and start the GUI up, as far as I can tell with it's location this is probably all main actually has to do for the project.

Pom.xml was updated because without specifying the compiler version it was unable to run with the current version of Java.