Finalized the programming part
Edited the GUI to throw errors for file writing exceptions, as well as connecting it to the algorithm.

Added a savelist component that actually runs the algorithm and saves the output of the algorithm to a text file called groupScheduling in the data component. This way we meet the requirment of outputing the result of the program

Removed the testing code from main and

Updated Group to be able to handle groups of differing sizes so the size input actually works

Finally removed any imports in java files that where unused